requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
python-slugify
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
